solving errors in user APIs
