+++ server/routes-controllers/api-routes.js
@@ -1,31 +1,8 @@
-// let db = require("../../mongo-database/index");
-
-// Initialize express router
-let router = require("express").Router();
-
-// Set default API response
-router.get("/", function (req, res) {
-  res.json({
-    status: "API Its Working",
-    message: "Welcome to RESTHub crafted with love!",
-  });
-});
-
-// Import user controller
-var userController = require("./userController");
-// user routes
-router.route("/users").get(userController.index).post(userController.new);
-router
-  .route("/users/:user_id")
-  .get(userController.view)
-  .patch(userController.update)
-  .put(userController.update)
-  .delete(userController.delete);
-// Export API routes
-module.exports = router;
+// // let db = require("../../mongo-database/index");
 
 // // Initialize express router
 // let router = require("express").Router();
+
 // // Set default API response
 // router.get("/", function (req, res) {
 //   res.json({
@@ -33,5 +10,28 @@ module.exports = router;
 //     message: "Welcome to RESTHub crafted with love!",
 //   });
 // });
+
+// // Import user controller
+// var userController = require("./userController");
+// // user routes
+// router.route("/users").get(userController.index).post(userController.new);
+// router
+//   .route("/users/:user_id")
+//   .get(userController.viewUser)
+//   .patch(userController.update)
+//   .put(userController.update)
+//   .delete(userController.delete);
 // // Export API routes
 // module.exports = router;
+
+// // // Initialize express router
+// // let router = require("express").Router();
+// // // Set default API response
+// // router.get("/", function (req, res) {
+// //   res.json({
+// //     status: "API Its Working",
+// //     message: "Welcome to RESTHub crafted with love!",
+// //   });
+// // });
+// // // Export API routes
+// // module.exports = router;
