+++ server/routes/api/index.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 
-router.use("/user", require("./user"));
+router.use("/users", require("./user"));
 // router.use("/resposes", require("./responses"));
 
 module.exports = router;
