+++ server/routes/api/user/index.js
@@ -3,11 +3,11 @@ const Controller = require("../../../controllers");
 
 router.get("/:id", Controller.UserController.getUser);
 
-router.post("/", Controller.UserController.addUser);
+router.post("/add", Controller.UserController.addUser);
 
 router.delete("/:id", Controller.UserController.deleteUser);
 
-router.put("/:id", Controller.UserController.updateUser);
+router.get("/getAll", Controller.UserController.viewUser);
 
 router.post("/login", Controller.AuthController.login);
